BEFORE 1st COMPETITION
Everything works. The shooter encoder is nonfunctional, and a test seeks
shooter encoder values. If no values are found, the shooter is powered
by voltage instead of Bang Bang. Autonomi are all good. Two Controllers
are used instead of One, with driving on the former and shoting/winch on
the latter.